requirements.txt
Django==1.8
psycopg2
sorl-thumbnail
Pillow